Incremental Co-operative Rebalancing Support for HDFS Connector
"TopicPartitionWriters" Map is being cleared on every close call, which is correct if we consider the stop the world scenario. But in Incremental mode there can be case that only few of the partitions have been revoked and task has some partitions with it to read from, In that scenario it will be throwing NullPointerException at the following part of the code.

As a consequence the add or the removal of an instance create a high pressure on all instances as they need to recover from last commited file for all the task of the cluster

Applying unmerged commit from original project : confluentinc#712